// ProBank/src/app/Models/Kunde.ts
export class Kunde {
    kundeId: number;
    fornavn: string;
    efternavn: string;
    email: string;
    telefonnummer: string;
    dob: Date;
    profilTekst: string;
    proStatus: boolean;
    sendtAnsoegning: boolean;
    godkendtArb01: number;
    godkendtArb02: number;
    godkendtArb03: number;
}